session 3 demo 2/requirements.txt
streamlit==1.32.0
langchain==0.1.13
langchain-community==0.0.29
langchain-google-genai==0.0.11
python-dotenv==1.0.1
pypdf==4.1.0
docx2txt==0.8
chromadb==0.4.24
google-generativeai==0.4.1
